requirements.txt
opencv-python==3.2.0.6
pytorch==1.1.0